+++ cli/src/run_wasm.rs
@@ -1,18 +1,29 @@
-use std::{path::PathBuf, str::FromStr};
+use std::{
+  path::{Path, PathBuf},
+  str::FromStr,
+  thread,
+  time::Duration,
+};
 
 use anyhow::Result;
 use clap::{command, CommandFactory, FromArgMatches, Parser};
+use notify_debouncer_mini::{
+  new_debouncer,
+  notify::{RecommendedWatcher, *},
+  DebounceEventResult, Debouncer,
+};
 
 use crate::{
   program_check::{check_all_programs, Program},
   CliCommand,
 };
 
+const WATCH_DEBOUNCE_GAP: Duration = Duration::from_secs(2);
+
 pub fn run_wasm() -> Box<dyn CliCommand> { Box::new(RunWasm {}) }
 
 struct RunWasm {}
-
-#[derive(Parser, Debug)]
+#[derive(Parser, Debug, Clone)]
 #[command(name = "run-wasm")]
 /// run as web wasm
 struct Wasm {
@@ -41,35 +52,70 @@ struct Wasm {
   template: Option<PathBuf>,
 }
 
-impl CliCommand for RunWasm {
-  fn name(&self) -> &str { "run-wasm" }
+impl Wasm {
+  fn root_path(&self) -> Result<PathBuf> {
+    let root = PathBuf::from_str(env!("CARGO_WORKSPACE_DIR"))?;
+    Ok(root)
+  }
 
-  fn command(&self) -> clap::Command { Wasm::command() }
-
-  fn exec(&self, args: &clap::ArgMatches) -> Result<()> {
-    let args = Wasm::from_arg_matches(args)?;
-
-    let mut dependencies =
-      vec![Program { crate_name: "wasm-bindgen-cli", binary_name: "wasm-bindgen" }];
-
-    if !args.no_server {
-      dependencies
-        .push(Program { crate_name: "simple-http-server", binary_name: "simple-http-server" });
-    }
-    check_all_programs(&dependencies)?;
-
-    let root_path = PathBuf::from_str(env!("CARGO_WORKSPACE_DIR"))?;
-    let package = args.package;
-    let name = args.name.unwrap_or("web_wasm".to_string());
-
-    let out_dir = args
+  fn out_dir(&self) -> PathBuf {
+    let out_dir = self
       .out_dir
+      .clone()
       .unwrap_or(PathBuf::from("./target/wasm"));
-    let output =
-      if out_dir.is_relative() { root_path.clone().join(&out_dir) } else { out_dir.clone() };
 
+    if out_dir.is_relative() {
+      let root_path = PathBuf::from_str(env!("CARGO_WORKSPACE_DIR")).unwrap();
+      root_path.clone().join(&out_dir)
+    } else {
+      out_dir.clone()
+    }
+  }
+
+  fn output_name(&self) -> String {
+    self
+      .name
+      .clone()
+      .unwrap_or("web_wasm".to_string())
+  }
+
+  fn auto_rebuild(&self) -> Debouncer<RecommendedWatcher> {
+    let root_path = self.root_path().unwrap();
+    let ignore_file = root_path.join(".gitignore");
+    let this = self.clone();
+    let mut debouncer =
+      new_debouncer(WATCH_DEBOUNCE_GAP, move |res: DebounceEventResult| match res {
+        Ok(events) => {
+          let ignore = gitignore::File::new(Path::new(&ignore_file));
+          let need_rebuild = events.iter().any(|e| {
+            if let Ok(ignore) = &ignore {
+              return !ignore.is_excluded(&e.path).unwrap_or(false);
+            }
+            true
+          });
+          if need_rebuild {
+            let _ = this.wasm_build();
+          }
+        }
+        Err(e) => println!("Error {:?}", e),
+      })
+      .unwrap();
+
+    // Add a path to be watched. All files and directories at that path and
+    // below will be monitored for changes.
+    debouncer
+      .watcher()
+      .watch(&root_path, RecursiveMode::Recursive)
+      .unwrap();
+    debouncer
+  }
+
+  fn wasm_build(&self) -> Result<()> {
     let shell = xshell::Shell::new()?;
-    let release_flg = if args.release { Some("--release") } else { None };
+    let root_path = self.root_path()?;
+    let output = self.out_dir();
+    let release_flg = if self.release { Some("--release") } else { None };
+    let package = self.package.clone();
 
     xshell::cmd!(
       shell,
@@ -79,21 +125,23 @@ impl CliCommand for RunWasm {
     .run()?;
 
     shell.change_dir(env!("CARGO_WORKSPACE_DIR"));
-    let target_path = if args.release {
+    let target_path = if self.release {
       "target/wasm32-unknown-unknown/release"
     } else {
       "target/wasm32-unknown-unknown/debug"
     };
 
+    let package = self.package.clone();
+    let name = self.output_name();
     xshell::cmd!(
       shell,
       "wasm-bindgen {target_path}/{package}.wasm --target web
-      --no-typescript --out-dir {output} --out-name {name}"
+    --no-typescript --out-dir {output} --out-name {name}"
     )
     .quiet()
     .run()?;
 
-    if let Some(mut path) = args.template.clone() {
+    if let Some(mut path) = self.template.clone() {
       if path.is_relative() {
         path = root_path.clone().join(path);
       }
@@ -114,15 +162,49 @@ impl CliCommand for RunWasm {
         )?;
       }
     }
+    Ok(())
+  }
 
-    if !args.no_server {
+  fn server(&self) -> Result<()> {
+    let root_path = self.root_path()?;
+    let _watcher = self.auto_rebuild();
+    let out_dir = self.out_dir();
+    let handle = thread::Builder::new().spawn(move || {
+      let shell = xshell::Shell::new().unwrap();
       shell.change_dir(root_path);
-      xshell::cmd!(
+      let _ = xshell::cmd!(
         shell,
-        "simple-http-server {out_dir} -c wasm,html,js -i --coep --coop --nocache"
+        "simple-http-server {out_dir} -c wasm,html,js -i --coep --coop
+          --nocache"
       )
       .quiet()
-      .run()?;
+      .run();
+    })?;
+
+    handle.join().unwrap();
+    Ok(())
+  }
+}
+
+impl CliCommand for RunWasm {
+  fn name(&self) -> &str { "run-wasm" }
+
+  fn command(&self) -> clap::Command { Wasm::command() }
+
+  fn exec(&self, args: &clap::ArgMatches) -> Result<()> {
+    let args = Wasm::from_arg_matches(args)?;
+
+    let mut dependencies =
+      vec![Program { crate_name: "wasm-bindgen-cli", binary_name: "wasm-bindgen" }];
+    if !args.no_server {
+      dependencies
+        .push(Program { crate_name: "simple-http-server", binary_name: "simple-http-server" });
+    }
+    check_all_programs(&dependencies)?;
+
+    args.wasm_build()?;
+    if !args.no_server {
+      args.server()?;
     }
 
     Ok(())